帯のしきい値を 1.097 → 1.15 に引き上げ（PQ_MIN_GH）
band_howto.md の検証にもとづき、本番の帯のしきい値を上げる。

  p/q>  探索2026/02-09  独立2025/05-2026/01  通し17か月
  1.097   99.2%           99.3%            99.3% [93,106] P=0.43  -2,570円
  1.150  102.1%          111.8%           107.2% [97,117] P=0.91 +10,830円

・両期間で単調に上がり 1.15 で頭打ち。1.20 で探索側が崩れる。
  1.13〜1.16 はどれも両期間100%超の平らな範囲で、その真ん中を取った。
・切り落とす側(1.097<p/q<=1.15)は探索97.1%/独立90.1%で、捨てる向きが正しい。
・点数は 3,541 → 1,508点（買った日あたり 7.3点 → 3.5点）。通知は約半分。
・穴側の ANA_PQ_MIN は 1.097 のまま据え置き（独立期間で87.5%、再現せず）。
  これまで PQ_MIN_GH と同値だったのでコメントを直した。
・波・風の足切りは現状維持。効いている確認は取れなかったが外す理由も無い。

あわせて ana_dump.py の日付を float64 に修正。float32 では8桁の日付が
丸められ（20260131→20260132）、期間の境界2日が両期間から落ちていた。
測り直しても結論は変わらない（差は0.5pt未満）。

selftest.py / select_rule.py の自己テスト / pytest 45件 すべて通過。

+++ ana_dump.py
@@ -43,9 +43,9 @@ assert m2 is not None and m3 is not None
 races = bt.collect(args)
 print(f"レース {len(races):,}", flush=True)
 
-FIRST1 = np.array([int(c[0]) for c in F.COMBOS], dtype=np.float32)
-SEC1 = np.array([int(c[2]) for c in F.COMBOS], dtype=np.float32)
-THI1 = np.array([int(c[4]) for c in F.COMBOS], dtype=np.float32)
+FIRST1 = np.array([int(c[0]) for c in F.COMBOS], dtype=np.float64)
+SEC1 = np.array([int(c[2]) for c in F.COMBOS], dtype=np.float64)
+THI1 = np.array([int(c[4]) for c in F.COMBOS], dtype=np.float64)
 NAMES = ["date", "race", "jcd", "rno", "combo", "first", "second", "third",
          "odds", "q", "p", "hit", "rmin", "band", "wave", "wind", "day_no",
          "is_final", "p1_first", "q1_first", "p_base", "p_g", "ok", "n_days"]
@@ -69,15 +69,17 @@ for rno_, (d, lanes, mt, od, q, q1, hit) in enumerate(races):
     if len(i) == 0:
         continue
     n = len(i)
-    one = lambda v: np.full(n, v, dtype=np.float32)
+    # ★float32 で持つと 8桁の日付が丸められる（20260131 → 20260132）。
+    #   日付で期間を切るので、ここは float64 でなければならない。
+    one = lambda v: np.full(n, v, dtype=np.float64)
     chunks.append(np.stack([
-        one(d), one(rno_), one(mt["jcd"]), one(mt["rno"]), i.astype(np.float32),
+        one(d), one(rno_), one(mt["jcd"]), one(mt["rno"]), i.astype(np.float64),
         FIRST1[i], SEC1[i], THI1[i], od[i], q[i], p[i], hitv[i], one(rmin),
         band[i], one(mt["wave"] if mt["wave"] is not None else -1),
         one(mt["wind"] if mt["wind"] is not None else -1),
         one(mt["day_no"] or -1), one(mt["is_final"]),
         p1[FIRST1[i].astype(int) - 1], q1[FIRST1[i].astype(int) - 1],
-        pb[i], pg[i], one(ok), one(mt["n_days"] or -1)], axis=1).astype(np.float32))
+        pb[i], pg[i], one(ok), one(mt["n_days"] or -1)], axis=1))
     if (rno_ + 1) % 10000 == 0:
         print(f"  {rno_+1}/{len(races)}", flush=True)
